requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
httpx==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.20.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
cryptography==42.0.8
bcrypt==4.1.3
